@@ -1,3 +1,4 @@
+import 'package:loggy/loggy.dart';
 import 'package:moor/moor.dart';
 import 'package:potato_notes/data/database.dart';
 
@@ -11,25 +12,46 @@ class NoteHelper extends DatabaseAccessor<AppDatabase> with _$NoteHelperMixin {
 
   Future<List<Note>> listNotes(ReturnMode mode) async {
     switch (mode) {
-      case ReturnMode.TAG:
+      case ReturnMode.ALL:
+        return select(notes).get();
       case ReturnMode.NORMAL:
         return (select(notes)
-              ..where((table) => table.archived.not() & table.deleted.not()))
+              ..where((table) =>
+                  table.archived.not() &
+                  table.deleted.not() &
+                  table.id.contains("-synced").not()))
             .get();
       case ReturnMode.ARCHIVE:
         return (select(notes)
-              ..where((table) => table.archived & table.deleted.not()))
+              ..where((table) =>
+                  table.archived &
+                  table.deleted.not() &
+                  table.id.contains("-synced").not()))
             .get();
       case ReturnMode.TRASH:
         return (select(notes)
-              ..where((table) => table.archived.not() & table.deleted))
+              ..where((table) =>
+                  table.archived.not() &
+                  table.deleted &
+                  table.id.contains("-synced").not()))
+            .get();
+      case ReturnMode.SYNCED:
+        return (select(notes)..where((table) => table.id.contains("-synced")))
+            .get();
+      case ReturnMode.TAG:
+      case ReturnMode.LOCAL:
+        return (select(notes)
+              ..where((table) => table.id.contains("-synced").not()))
             .get();
       case ReturnMode.FAVOURITES:
         return (select(notes)
               ..where((table) =>
-                  table.starred & table.archived.not() & table.deleted.not()))
+                  table.starred &
+                  table.archived.not() &
+                  table.deleted.not() &
+                  table.id.contains("-synced").not()))
             .get();
-      case ReturnMode.ALL:
+
       default:
         return select(notes).get();
     }
@@ -39,26 +61,46 @@ class NoteHelper extends DatabaseAccessor<AppDatabase> with _$NoteHelperMixin {
     SimpleSelectStatement<$NotesTable, Note> selectQuery;
 
     switch (mode) {
-      case ReturnMode.TAG:
       case ReturnMode.ALL:
         selectQuery = select(notes);
         break;
       case ReturnMode.NORMAL:
         selectQuery = select(notes)
-          ..where((table) => table.archived.not() & table.deleted.not());
+          ..where((table) =>
+              table.archived.not() &
+              table.deleted.not() &
+              table.id.contains("-synced").not());
         break;
       case ReturnMode.ARCHIVE:
         selectQuery = select(notes)
-          ..where((table) => table.archived & table.deleted.not());
+          ..where((table) =>
+              table.archived &
+              table.deleted.not() &
+              table.id.contains("-synced").not());
         break;
       case ReturnMode.FAVOURITES:
         selectQuery = select(notes)
           ..where((table) =>
-              table.starred & table.archived.not() & table.deleted.not());
+              table.starred &
+              table.archived.not() &
+              table.deleted.not() &
+              table.id.contains("-synced").not());
         break;
       case ReturnMode.TRASH:
         selectQuery = select(notes)
-          ..where((table) => table.archived.not() & table.deleted);
+          ..where((table) =>
+              table.archived.not() &
+              table.deleted &
+              table.id.contains("-synced").not());
+        break;
+      case ReturnMode.SYNCED:
+        selectQuery = select(notes)
+          ..where((table) => table.id.contains("-synced"));
+        break;
+      case ReturnMode.TAG:
+      case ReturnMode.LOCAL:
+        selectQuery = select(notes)
+          ..where((table) => table.id.contains("-synced").not());
         break;
     }
 
@@ -70,10 +112,23 @@ class NoteHelper extends DatabaseAccessor<AppDatabase> with _$NoteHelperMixin {
         .watch();
   }
 
-  Future saveNote(Note note) =>
-      into(notes).insert(note, mode: InsertMode.replace);
+  Future<void> saveNote(Note note) {
+    Loggy.d(message: "The note id is: " + note.id);
+    return into(notes).insert(note, mode: InsertMode.replace);
+  }
 
-  Future deleteNote(Note note) => delete(notes).delete(note);
+  Future<void> deleteNote(Note note) {
+    Loggy.d(message: "The note id to delete: " + note.id);
+    return delete(notes).delete(note);
+  }
+
+  Future<void> deleteAllNotes() async {
+    List<Note> notes = await listNotes(ReturnMode.ALL);
+
+    notes.forEach((note) async {
+      await deleteNote(note);
+    });
+  }
 }
 
 class SearchQuery {
@@ -81,8 +136,11 @@ class SearchQuery {
   int _color;
   DateTime date;
   DateFilterMode dateMode;
+  List<String> tags = [];
+  bool onlyFavourites;
 
   int get color => _color ?? 0;
+
   set color(int value) {
     if (value == -1) {
       _color = 0;
@@ -96,6 +154,7 @@ class SearchQuery {
     int color,
     this.date,
     this.dateMode = DateFilterMode.ONLY,
+    this.onlyFavourites = false,
   }) : _color = color;
 }
 
@@ -105,11 +164,4 @@ enum DateFilterMode {
   ONLY,
 }
 
-enum ReturnMode {
-  ALL,
-  NORMAL,
-  ARCHIVE,
-  TRASH,
-  FAVOURITES,
-  TAG,
-}
+enum ReturnMode { ALL, NORMAL, ARCHIVE, TRASH, FAVOURITES, TAG, SYNCED, LOCAL }
