@@ -2,5 +2,8 @@ import 'package:moor/backends.dart';
 import 'package:moor/moor_web.dart';
 
 QueryExecutor constructDb({bool logStatements = false}) {
-  return WebDatabase('db', logStatements: logStatements);
+  return WebDatabase.withStorage(
+    MoorWebStorage.indexedDbIfSupported('notes'),
+    logStatements: logStatements,
+  );
 }
