@@ -1,38 +1,34 @@
 import 'dart:convert';
 
-import 'package:json_annotation/json_annotation.dart';
 import 'package:moor/moor.dart';
 
-part 'reminder_list.g.dart';
-
-@JsonSerializable()
-class ReminderList {
-  List<DateTime> reminders;
-
-  ReminderList(this.reminders);
-
-  factory ReminderList.fromJson(Map<String, dynamic> json) =>
-      _$ReminderListFromJson(json);
-
-  Map<String, dynamic> toJson() => _$ReminderListToJson(this);
-}
-
-class ReminderListConverter extends TypeConverter<ReminderList, String> {
+class ReminderListConverter extends TypeConverter<List<DateTime>, String> {
   const ReminderListConverter();
   @override
-  ReminderList mapToDart(String fromDb) {
+  List<DateTime> mapToDart(String fromDb) {
     if (fromDb == null) {
       return null;
     }
-    return ReminderList.fromJson(json.decode(fromDb) as Map<String, dynamic>);
+    List<dynamic> decoded = json.decode(fromDb);
+    return List.generate(
+      decoded.length,
+      (index) => DateTime.fromMillisecondsSinceEpoch(
+        json.decode(decoded[index]),
+      ),
+    );
   }
 
   @override
-  String mapToSql(ReminderList value) {
+  String mapToSql(List<DateTime> value) {
     if (value == null) {
       return null;
     }
 
-    return json.encode(value.toJson());
+    return json.encode(
+      List.generate(
+        value.length,
+        (index) => value[index].millisecondsSinceEpoch,
+      ),
+    );
   }
 }
