@@ -1,38 +1,24 @@
 import 'dart:convert';
 
-import 'package:json_annotation/json_annotation.dart';
 import 'package:moor/moor.dart';
 
-part 'tag_list.g.dart';
-
-@JsonSerializable()
-class TagList {
-  List<int> tagIds;
-
-  TagList(this.tagIds);
-
-  factory TagList.fromJson(Map<String, dynamic> json) =>
-      _$TagListFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TagListToJson(this);
-}
-
-class TagListConverter extends TypeConverter<TagList, String> {
+class TagListConverter extends TypeConverter<List<String>, String> {
   const TagListConverter();
   @override
-  TagList mapToDart(String fromDb) {
+  List<String> mapToDart(String fromDb) {
     if (fromDb == null) {
       return null;
     }
-    return TagList.fromJson(json.decode(fromDb) as Map<String, dynamic>);
+    List<dynamic> decoded = json.decode(fromDb);
+    return List.generate(decoded.length, (index) => decoded[index] as String);
   }
 
   @override
-  String mapToSql(TagList value) {
+  String mapToSql(List<String> value) {
     if (value == null) {
       return null;
     }
 
-    return json.encode(value.toJson());
+    return json.encode(value);
   }
 }
