@@ -1,21 +1,32 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:package_info/package_info.dart';
+import 'package:mobx/mobx.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:potato_notes/internal/device_info.dart';
 import 'package:potato_notes/internal/illustrations.dart';
 import 'package:potato_notes/internal/notification_payload.dart';
 import 'package:quick_actions/quick_actions.dart';
 
-class AppInfo {
-  AppInfo() {
+part 'app_info.g.dart';
+
+class AppInfo = _AppInfoBase with _$AppInfo;
+
+abstract class _AppInfoBase with Store {
+  static final EventChannel accentStreamChannel =
+      EventChannel('potato_notes_accents');
+
+  _AppInfoBase() {
     illustrations = Illustrations();
     loadData();
   }
 
+  Directory tempDirectory;
   Illustrations illustrations;
   FlutterLocalNotificationsPlugin notifications;
   QuickActions quickActions;
@@ -27,6 +38,18 @@ class AppInfo {
   Widget noFavouritesIllustration;
   Widget nothingFoundIllustration;
   Widget typeToSearchIllustration;
+
+  @observable
+  @protected
+  int accentDataValue = Colors.blue.value;
+
+  int get accentData => accentDataValue;
+
+  @observable
+  @protected
+  List<ActiveNotification> activeNotificationsValue = [];
+
+  List<ActiveNotification> get activeNotifications => activeNotificationsValue;
 
   void updateIllustrations(Brightness systemTheme) async {
     noNotesIllustration = await illustrations.noNotesIllustration(systemTheme);
@@ -47,8 +70,12 @@ class AppInfo {
     var initializationSettingsAndroid =
         AndroidInitializationSettings('notes_icon');
     var initializationSettingsIOS = IOSInitializationSettings();
+    var initializationSettingsMacOS = MacOSInitializationSettings();
     var initializationSettings = InitializationSettings(
-        initializationSettingsAndroid, initializationSettingsIOS);
+      android: initializationSettingsAndroid,
+      iOS: initializationSettingsIOS,
+      macOS: initializationSettingsMacOS,
+    );
     await notifications.initialize(initializationSettings,
         onSelectNotification: _handleNotificationTap);
   }
@@ -67,10 +94,36 @@ class AppInfo {
   }
 
   void loadData() async {
-    if (!kIsWeb) {
+    tempDirectory = await getTemporaryDirectory();
+
+    if (!DeviceInfo.isDesktopOrWeb) {
       _initNotifications();
+      packageInfo = await PackageInfo.fromPlatform();
     }
 
-    packageInfo = await PackageInfo.fromPlatform();
+    if (DeviceInfo.isAndroid) {
+      accentStreamChannel.receiveBroadcastStream().listen(updateAccent);
+    } else {
+      accentDataValue = -1;
+    }
+
+    if (DeviceInfo.isAndroid) {
+      pollForActiveNotifications();
+    }
+  }
+
+  @action
+  void updateAccent(dynamic event) {
+    accentDataValue = event as int;
+  }
+
+  @action
+  void pollForActiveNotifications() {
+    Timer.periodic(Duration(milliseconds: 500), (timer) async {
+      activeNotificationsValue = await notifications
+          .resolvePlatformSpecificImplementation<
+              AndroidFlutterLocalNotificationsPlugin>()
+          .getActiveNotifications();
+    });
   }
 }
