@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:potato_notes/internal/locale_strings.dart';
+import 'package:potato_notes/internal/locales/locale_strings.g.dart';
 
 class NoteColors {
   static List<ColorInfo> get colorList => [
