@@ -1,50 +1,124 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:local_auth/local_auth.dart';
+import 'package:mobx/mobx.dart';
 
-class DeviceInfo {
-  bool _canCheckBiometrics;
-  List<BiometricType> _availableBiometrics;
-  bool _canUseSystemAccent = true;
-  bool _isLandscape = false;
-  int _uiSizeFactor = 2;
-  UiType _uiType;
+part 'device_info.g.dart';
 
-  DeviceInfo() {
+class DeviceInfo extends _DeviceInfoBase with _$DeviceInfo {
+  static bool get isDesktopOrWeb {
+    if (kIsWeb) return true;
+
+    if (Platform.isLinux || Platform.isWindows || Platform.isMacOS) return true;
+
+    return false;
+  }
+
+  static bool get isDesktop {
+    if (kIsWeb) return false;
+
+    if (Platform.isLinux || Platform.isWindows || Platform.isMacOS) return true;
+
+    return false;
+  }
+
+  static bool get isAndroid {
+    if (kIsWeb) return false;
+
+    if (Platform.isAndroid) return true;
+
+    return false;
+  }
+
+  static bool get isMacOS {
+    if (kIsWeb) return false;
+
+    if (Platform.isMacOS) return true;
+
+    return false;
+  }
+}
+
+abstract class _DeviceInfoBase with Store {
+  @observable
+  @protected
+  bool canCheckBiometricsValue;
+
+  @observable
+  @protected
+  List<BiometricType> availableBiometricsValue;
+
+  @observable
+  @protected
+  bool canUseSystemAccentValue = true;
+
+  @observable
+  @protected
+  bool isLandscapeValue = false;
+
+  @observable
+  @protected
+  int uiSizeFactorValue = 2;
+
+  @observable
+  @protected
+  UiType uiTypeValue;
+
+  _DeviceInfoBase() {
     _loadInitialData();
   }
 
-  bool get canCheckBiometrics => _canCheckBiometrics;
-  List<BiometricType> get availableBiometrics => _availableBiometrics;
-  bool get canUseSystemAccent => _canUseSystemAccent;
-  bool get isLandscape => _isLandscape;
-  int get uiSizeFactor => _uiSizeFactor;
-  UiType get uiType => _uiType;
+  bool get canCheckBiometrics => canCheckBiometricsValue;
+  List<BiometricType> get availableBiometrics => availableBiometricsValue;
+  bool get canUseSystemAccent => canUseSystemAccentValue;
+  bool get isLandscape => isLandscapeValue;
+  int get uiSizeFactor => uiSizeFactorValue;
+  UiType get uiType => uiTypeValue;
 
-  void _loadInitialData() async {
-    _canCheckBiometrics = await LocalAuthentication().canCheckBiometrics;
-    _availableBiometrics = await LocalAuthentication().getAvailableBiometrics();
+  @action
+  Future<void> _loadInitialData() async {
+    if (!DeviceInfo.isDesktopOrWeb) {
+      canCheckBiometricsValue = await LocalAuthentication().canCheckBiometrics;
+      availableBiometricsValue =
+          await LocalAuthentication().getAvailableBiometrics();
+    } else {
+      canCheckBiometricsValue = false;
+      availableBiometricsValue = [];
+    }
   }
 
+  @action
   void updateDeviceInfo(MediaQueryData mq, bool canUseSystemAccent) {
-    _canUseSystemAccent = canUseSystemAccent;
-    _isLandscape = mq.orientation == Orientation.landscape;
+    canUseSystemAccentValue = canUseSystemAccent;
+    isLandscapeValue = mq.orientation == Orientation.landscape;
     double width = mq.size.width;
 
-    if (width >= 1280) {
-      _uiSizeFactor = 5;
-      _uiType = UiType.DESKTOP;
+    if (width >= 1920) {
+      uiSizeFactorValue = 8;
+      uiTypeValue = UiType.DESKTOP;
+    } else if (width >= 1600) {
+      uiSizeFactorValue = 7;
+      uiTypeValue = UiType.DESKTOP;
+    } else if (width >= 1460) {
+      uiSizeFactorValue = 6;
+      uiTypeValue = UiType.DESKTOP;
+    } else if (width >= 1280) {
+      uiSizeFactorValue = 5;
+      uiTypeValue = UiType.DESKTOP;
     } else if (width >= 900) {
-      _uiSizeFactor = 4;
-      _uiType = UiType.LARGE_TABLET;
+      uiSizeFactorValue = 4;
+      uiTypeValue = UiType.LARGE_TABLET;
     } else if (width >= 600) {
-      _uiSizeFactor = 3;
-      _uiType = UiType.TABLET;
+      uiSizeFactorValue = 3;
+      uiTypeValue = UiType.TABLET;
     } else if (width >= 360) {
-      _uiSizeFactor = 2;
-      _uiType = UiType.PHONE;
+      uiSizeFactorValue = 2;
+      uiTypeValue = UiType.PHONE;
     } else {
-      _uiSizeFactor = 1;
-      _uiType = UiType.PHONE;
+      uiSizeFactorValue = 1;
+      uiTypeValue = UiType.PHONE;
     }
   }
 }
