@@ -1,12 +1,15 @@
 import 'package:potato_notes/data/dao/note_helper.dart';
+import 'package:potato_notes/data/dao/tag_helper.dart';
 import 'package:potato_notes/internal/app_info.dart';
 import 'package:potato_notes/internal/device_info.dart';
 import 'package:potato_notes/internal/preferences.dart';
 
-AppInfo appInfo;
+final AppInfo appInfo = AppInfo();
 
-DeviceInfo deviceInfo;
+final DeviceInfo deviceInfo = DeviceInfo();
 
-Preferences prefs;
+final Preferences prefs = Preferences();
 
 NoteHelper helper;
+
+TagHelper tagHelper;
