@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:potato_notes/internal/locale_strings.dart';
+import 'package:potato_notes/internal/locales/locale_strings.g.dart';
+import 'package:potato_notes/widget/dependent_scaffold.dart';
 
 class SearchPage extends StatefulWidget {
   final CustomSearchDelegate delegate;
@@ -20,6 +21,7 @@ class _SearchPageState<T> extends State<SearchPage> {
     super.initState();
     widget.delegate._queryTextController.addListener(_onQueryChanged);
     widget.delegate._focusNode = focusNode;
+    widget.delegate._setState = setState;
   }
 
   @override
@@ -44,46 +46,26 @@ class _SearchPageState<T> extends State<SearchPage> {
 
   @override
   Widget build(BuildContext context) {
-    assert(debugCheckHasMaterialLocalizations(context));
-    final ThemeData theme = Theme.of(context);
-    final String searchFieldLabel = widget.delegate.searchFieldLabel ??
-        MaterialLocalizations.of(context).searchFieldLabel;
     Widget body = widget.delegate.buildResults(context);
-    String routeName;
-    switch (theme.platform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        routeName = '';
-        break;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        routeName = searchFieldLabel;
-    }
 
-    return Semantics(
-      explicitChildNodes: true,
-      scopesRoute: true,
-      namesRoute: true,
-      label: routeName,
-      child: Scaffold(
-        appBar: AppBar(
-          title: Padding(
-            padding: EdgeInsets.symmetric(horizontal: 0),
-            child: TextField(
-              controller: widget.delegate._queryTextController,
-              focusNode: focusNode,
-              decoration: InputDecoration.collapsed(
-                hintText: LocaleStrings.searchPage.textboxHint,
-              ),
-              onChanged: (value) => _onSearchBodyChanged(),
+    return DependentScaffold(
+      appBar: AppBar(
+        title: Padding(
+          padding: EdgeInsets.symmetric(horizontal: 0),
+          child: TextField(
+            controller: widget.delegate._queryTextController,
+            focusNode: focusNode,
+            decoration: InputDecoration.collapsed(
+              hintText: LocaleStrings.searchPage.textboxHint,
             ),
+            autofocus: true,
+            onChanged: (value) => _onSearchBodyChanged(),
           ),
-          actions: widget.delegate.buildActions(context),
         ),
-        body: body,
+        actions: widget.delegate.buildActions(context),
       ),
+      useAppBarAsSecondary: true,
+      body: body,
     );
   }
 }
@@ -124,4 +106,8 @@ abstract class CustomSearchDelegate<T> {
 
   final ProxyAnimation _proxyAnimation =
       ProxyAnimation(kAlwaysDismissedAnimation);
+
+  void Function(VoidCallback) _setState;
+
+  void Function(VoidCallback) get setState => _setState;
 }
