@@ -1,8 +1,8 @@
 import 'package:animations/animations.dart';
 import 'package:back_button_interceptor/back_button_interceptor.dart';
 import 'package:flutter/material.dart';
-import 'package:potato_notes/internal/locale_strings.dart';
 import 'package:potato_notes/internal/providers.dart';
+import 'package:potato_notes/internal/locales/locale_strings.g.dart';
 import 'package:potato_notes/routes/setup/finish_page.dart';
 import 'package:potato_notes/routes/setup/basic_customization_page.dart';
 import 'package:potato_notes/routes/setup/welcome_page.dart';
@@ -37,6 +37,7 @@ class _SetupPagetate extends State<SetupPage> {
 
   @override
   Widget build(BuildContext context) {
+    TextDirection textDirection = Directionality.of(context);
     String buttonText;
 
     if (pageIndex == 0) {
@@ -50,7 +51,9 @@ class _SetupPagetate extends State<SetupPage> {
       child: Scaffold(
         body: GestureDetector(
           onHorizontalDragEnd: (details) {
-            double sign = details.primaryVelocity.sign;
+            double sign = textDirection == TextDirection.rtl
+                ? -details.primaryVelocity.sign
+                : details.primaryVelocity.sign;
 
             if (details.primaryVelocity.abs() > 320) {
               if (sign == 1) {
@@ -65,7 +68,7 @@ class _SetupPagetate extends State<SetupPage> {
             }
           },
           child: PageTransitionSwitcher(
-            reverse: reverse,
+            reverse: textDirection == TextDirection.rtl ? !reverse : reverse,
             transitionBuilder: (
               child,
               primaryAnimation,
