@@ -1,13 +1,14 @@
 import 'dart:io';
 
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_blurhash/flutter_blurhash.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
+import 'package:potato_notes/data/model/saved_image.dart';
 import 'package:potato_notes/internal/providers.dart';
 import 'package:potato_notes/internal/utils.dart';
 
 class NoteViewImages extends StatefulWidget {
-  final List<Uri> images;
+  final List<SavedImage> images;
   final double borderRadius;
   final bool showPlusImages;
   final int numPlusImages;
@@ -55,24 +56,20 @@ class _NoteViewImagesState extends State<NoteViewImages> {
                   ? kMaxImageCount
                   : widget.images.length,
               itemBuilder: (context, index) {
+                SavedImage savedImage = widget.images[index];
                 ImageProvider image;
-                String scheme = widget.images[index].scheme;
-
-                if (scheme.startsWith("http")) {
-                  image = CachedNetworkImageProvider(
-                      widget.images[index].toString());
+                if (savedImage.existsLocally) {
+                  image = FileImage(File(savedImage.path));
                 } else {
-                  image = FileImage(File(widget.images[index].path));
+                  image = BlurHashImage(savedImage.blurHash);
                 }
-
                 return Stack(
                   children: [
                     SizedBox.expand(
                       child: Image(
-                        image: image,
-                        fit: BoxFit.cover,
-                        gaplessPlayback: true,
-                      ),
+                          image: image,
+                          fit: BoxFit.cover,
+                          gaplessPlayback: true),
                     ),
                     SizedBox.expand(
                       child: Visibility(
@@ -103,6 +100,55 @@ class _NoteViewImagesState extends State<NoteViewImages> {
                     ),
                   ],
                 );
+                //ImageProvider image;
+                /*String scheme = widget.images[index].uri.scheme;
+
+                if (scheme.startsWith("http")) {
+                  image = CachedNetworkImageProvider(
+                      widget.images[index].toString());
+                } else {*/
+                /*if (widget.images[index].uri != null) {
+
+                }*/
+                /*if (!ImageService.imageCached(widget.images[index])) {
+                  imageService
+                      .downloadImage(widget.images[index])
+                      .then((path) => {
+                            if (path != "") {setState(() => {})}
+                          });
+                }*/
+
+                /*
+                Image(
+                  image: image
+                  fit: BoxFit.cover,
+                  gaplessPlayback: true,
+                );
+                SavedImage savedImage = widget.images[index];
+                if(savedImage.existsLocally){
+
+                }
+                final Widget image = OctoImage(
+                  image: CachedNetworkImageProvider(
+                    () async {
+                      try {
+                        String url =
+                            await ImageController.getDownloadUrlFromSync(
+                                widget.images[index].hash);
+                        return url;
+                      } catch (e) {
+                        return "";
+                      }
+                    },
+                    cacheKey: widget.images[index].hash,
+                  ),
+                  placeholderBuilder:
+                      OctoPlaceholder.blurHash(widget.images[index].blurHash),
+                  fit: BoxFit.cover,
+                  gaplessPlayback: false,
+                  fadeInDuration: const Duration(milliseconds: 1000),
+                  fadeOutDuration: const Duration(milliseconds: 500),
+                );*/
               },
               staggeredTileBuilder: (index) {
                 int crossAxisExtent = 1;
