@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:outline_material_icons/outline_material_icons.dart';
 import 'package:potato_notes/data/dao/note_helper.dart';
 import 'package:potato_notes/internal/colors.dart';
-import 'package:potato_notes/internal/locale_strings.dart';
 import 'package:potato_notes/internal/utils.dart';
+import 'package:potato_notes/internal/locales/locale_strings.g.dart';
+import 'package:potato_notes/routes/search_page.dart';
 import 'package:potato_notes/widget/date_selector.dart';
+import 'package:potato_notes/widget/tag_search_delegate.dart';
 
 import 'note_color_selector.dart';
 
@@ -39,8 +40,19 @@ class _QueryFiltersState extends State<QueryFilters> {
               widget.filterChangedCallback();
           },
         ),
+        SwitchListTile(
+          secondary: Icon(Icons.favorite_border),
+          title: Text("Favourites only"),
+          value: widget.query.onlyFavourites,
+          activeColor: Theme.of(context).accentColor,
+          onChanged: (value) {
+            setState(() => widget.query.onlyFavourites = value);
+            if (widget.filterChangedCallback != null)
+              widget.filterChangedCallback();
+          },
+        ),
         ListTile(
-          leading: Icon(OMIcons.colorLens),
+          leading: Icon(Icons.color_lens_outlined),
           title: Text(LocaleStrings.common.colorFilter),
           trailing: Icon(
             Icons.brightness_1,
@@ -70,7 +82,7 @@ class _QueryFiltersState extends State<QueryFilters> {
           },
         ),
         ListTile(
-          leading: Icon(OMIcons.dateRange),
+          leading: Icon(Icons.date_range_outlined),
           title: Text(LocaleStrings.common.dateFilter),
           subtitle: widget.query.date != null
               ? Text(DateFormat("EEEE d MMM yyyy").format(widget.query.date) +
@@ -101,6 +113,29 @@ class _QueryFiltersState extends State<QueryFilters> {
                           widget.filterChangedCallback();
                       },
                     ));
+          },
+        ),
+        ListTile(
+          leading: Icon(Icons.label_outline),
+          title: Text("Tags"),
+          trailing: Text(
+            widget.query.tags.isNotEmpty
+                ? "${widget.query.tags.length} selected"
+                : "",
+          ),
+          onTap: () async {
+            await Utils.showSecondaryRoute(
+              context,
+              SearchPage(
+                delegate: TagSearchDelegate(
+                  widget.query.tags,
+                  onChanged: () => setState(() {}),
+                ),
+              ),
+            );
+
+            if (widget.filterChangedCallback != null)
+              widget.filterChangedCallback();
           },
         ),
       ],
